Module esp32c6::efuse::rd_repeat_data0

BLOCK0 data register 1.

Structs

RD_REPEAT_DATA0_SPEC

BLOCK0 data register 1.

Type Aliases

DIS_CAN_R

Field DIS_CAN reader - Represents whether TWAI function is disabled or enabled. 1: disabled. 0: enabled.

DIS_DOWNLOAD_ICACHE_R

Field DIS_DOWNLOAD_ICACHE reader - Represents whether icache is disabled or enabled in Download mode. 1: disabled. 0: enabled.

DIS_DOWNLOAD_MANUAL_ENCRYPT_R

Field DIS_DOWNLOAD_MANUAL_ENCRYPT reader - Represents whether flash encrypt function is disabled or enabled(except in SPI boot mode). 1: disabled. 0: enabled.

DIS_FORCE_DOWNLOAD_R

Field DIS_FORCE_DOWNLOAD reader - Represents whether the function that forces chip into download mode is disabled or enabled. 1: disabled. 0: enabled.

DIS_ICACHE_R

Field DIS_ICACHE reader - Represents whether icache is disabled or enabled. 1: disabled. 0: enabled.

DIS_PAD_JTAG_R

Field DIS_PAD_JTAG reader - Represents whether JTAG is disabled in the hard way(permanently). 1: disabled. 0: enabled.

DIS_USB_JTAG_R

Field DIS_USB_JTAG reader - Represents whether the function of usb switch to jtag is disabled or enabled. 1: disabled. 0: enabled.

DIS_USB_SERIAL_JTAG_R

Field DIS_USB_SERIAL_JTAG reader - Represents whether USB-Serial-JTAG is disabled or enabled. 1: disabled. 0: enabled.

JTAG_SEL_ENABLE_R

Field JTAG_SEL_ENABLE reader - Represents whether the selection between usb_to_jtag and pad_to_jtag through strapping gpio15 when both EFUSE_DIS_PAD_JTAG and EFUSE_DIS_USB_JTAG are equal to 0 is enabled or disabled. 1: enabled. 0: disabled.

R

Register RD_REPEAT_DATA0 reader

RD_DIS_R

Field RD_DIS reader - Represents whether reading of individual eFuse block(block4~block10) is disabled or enabled. 1: disabled. 0: enabled.

RPT4_RESERVED0_0_R

Field RPT4_RESERVED0_0 reader - Reserved.

RPT4_RESERVED0_1_R

Field RPT4_RESERVED0_1 reader - Reserved.

RPT4_RESERVED0_2_R

Field RPT4_RESERVED0_2 reader - Reserved.

SOFT_DIS_JTAG_R

Field SOFT_DIS_JTAG reader - Represents whether JTAG is disabled in soft way. Odd number: disabled. Even number: enabled.

SPI_DOWNLOAD_MSPI_DIS_R

Field SPI_DOWNLOAD_MSPI_DIS reader - Represents whether SPI0 controller during boot_mode_download is disabled or enabled. 1: disabled. 0: enabled.

SWAP_UART_SDIO_EN_R

Field SWAP_UART_SDIO_EN reader - Represents whether pad of uart and sdio is swapped or not. 1: swapped. 0: not swapped.

USB_DREFH_R

Field USB_DREFH reader - Represents the single-end input threhold vrefh, 1.76 V to 2 V with step of 80 mV.

USB_DREFL_R

Field USB_DREFL reader - Represents the single-end input threhold vrefl, 1.76 V to 2 V with step of 80 mV.

USB_EXCHG_PINS_R

Field USB_EXCHG_PINS reader - Represents whether the D+ and D- pins is exchanged. 1: exchanged. 0: not exchanged.

VDD_SPI_AS_GPIO_R

Field VDD_SPI_AS_GPIO reader - Represents whether vdd spi pin is functioned as gpio. 1: functioned. 0: not functioned.