Module esp32c6::efuse::rd_repeat_data3

BLOCK0 data register 4.

Structs

RD_REPEAT_DATA3_SPEC

BLOCK0 data register 4.

Type Aliases

DIS_DIRECT_BOOT_R

Field DIS_DIRECT_BOOT reader - Represents whether direct boot mode is disabled or enabled. 1: disabled. 0: enabled.

DIS_DOWNLOAD_MODE_R

Field DIS_DOWNLOAD_MODE reader - Represents whether Download mode is disabled or enabled. 1: disabled. 0: enabled.

DIS_USB_PRINT_R

Field DIS_USB_PRINT reader - Represents whether print from USB-Serial-JTAG is disabled or enabled. 1: disabled. 0: enabled.

DIS_USB_SERIAL_JTAG_DOWNLOAD_MODE_R

Field DIS_USB_SERIAL_JTAG_DOWNLOAD_MODE reader - Represents whether the USB-Serial-JTAG download function is disabled or enabled. 1: disabled. 0: enabled.

ENABLE_SECURITY_DOWNLOAD_R

Field ENABLE_SECURITY_DOWNLOAD reader - Represents whether security download is enabled or disabled. 1: enabled. 0: disabled.

FORCE_SEND_RESUME_R

Field FORCE_SEND_RESUME reader - Represents whether ROM code is forced to send a resume command during SPI boot. 1: forced. 0:not forced.

R

Register RD_REPEAT_DATA3 reader

RPT4_RESERVED3_0_R

Field RPT4_RESERVED3_0 reader - Reserved.

RPT4_RESERVED3_1_R

Field RPT4_RESERVED3_1 reader - Reserved.

RPT4_RESERVED3_2_R

Field RPT4_RESERVED3_2 reader - Reserved.

RPT4_RESERVED3_3_R

Field RPT4_RESERVED3_3 reader - Reserved.

RPT4_RESERVED3_4_R

Field RPT4_RESERVED3_4 reader - Reserved.

RPT4_RESERVED3_5_R

Field RPT4_RESERVED3_5 reader - Reserved.

SECURE_BOOT_DISABLE_FAST_WAKE_R

Field SECURE_BOOT_DISABLE_FAST_WAKE reader - Represents whether FAST VERIFY ON WAKE is disabled or enabled when Secure Boot is enabled. 1: disabled. 0: enabled.

SECURE_VERSION_R

Field SECURE_VERSION reader - Represents the version used by ESP-IDF anti-rollback feature.

UART_PRINT_CONTROL_R

Field UART_PRINT_CONTROL reader - Represents the type of UART printing. 00: force enable printing. 01: enable printing when GPIO8 is reset at low level. 10: enable printing when GPIO8 is reset at high level. 11: force disable printing.