Type Alias esp32c6::extmem::l1_cache_sync_preload_int_raw::W

pub type W = W<L1_CACHE_SYNC_PRELOAD_INT_RAW_SPEC>;

Register L1_CACHE_SYNC_PRELOAD_INT_RAW writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn l1_icache0_pld_done_int_raw( &mut self ) -> L1_ICACHE0_PLD_DONE_INT_RAW_W<'_, L1_CACHE_SYNC_PRELOAD_INT_RAW_SPEC>

Bit 0 - The raw bit of the interrupt that occurs only when L1-ICache0 preload-operation is done.

pub fn l1_icache1_pld_done_int_raw( &mut self ) -> L1_ICACHE1_PLD_DONE_INT_RAW_W<'_, L1_CACHE_SYNC_PRELOAD_INT_RAW_SPEC>

Bit 1 - The raw bit of the interrupt that occurs only when L1-ICache1 preload-operation is done.

pub fn l1_icache2_pld_done_int_raw( &mut self ) -> L1_ICACHE2_PLD_DONE_INT_RAW_W<'_, L1_CACHE_SYNC_PRELOAD_INT_RAW_SPEC>

Bit 2 - Reserved

pub fn l1_icache3_pld_done_int_raw( &mut self ) -> L1_ICACHE3_PLD_DONE_INT_RAW_W<'_, L1_CACHE_SYNC_PRELOAD_INT_RAW_SPEC>

Bit 3 - Reserved

pub fn l1_cache_pld_done_int_raw( &mut self ) -> L1_CACHE_PLD_DONE_INT_RAW_W<'_, L1_CACHE_SYNC_PRELOAD_INT_RAW_SPEC>

Bit 4 - The raw bit of the interrupt that occurs only when L1-Cache preload-operation is done.

pub fn cache_sync_done_int_raw( &mut self ) -> CACHE_SYNC_DONE_INT_RAW_W<'_, L1_CACHE_SYNC_PRELOAD_INT_RAW_SPEC>

Bit 6 - The raw bit of the interrupt that occurs only when Cache sync-operation is done.

pub fn l1_icache0_pld_err_int_raw( &mut self ) -> L1_ICACHE0_PLD_ERR_INT_RAW_W<'_, L1_CACHE_SYNC_PRELOAD_INT_RAW_SPEC>

Bit 7 - The raw bit of the interrupt that occurs only when L1-ICache0 preload-operation error occurs.

pub fn l1_icache1_pld_err_int_raw( &mut self ) -> L1_ICACHE1_PLD_ERR_INT_RAW_W<'_, L1_CACHE_SYNC_PRELOAD_INT_RAW_SPEC>

Bit 8 - The raw bit of the interrupt that occurs only when L1-ICache1 preload-operation error occurs.

pub fn l1_icache2_pld_err_int_raw( &mut self ) -> L1_ICACHE2_PLD_ERR_INT_RAW_W<'_, L1_CACHE_SYNC_PRELOAD_INT_RAW_SPEC>

Bit 9 - Reserved

pub fn l1_icache3_pld_err_int_raw( &mut self ) -> L1_ICACHE3_PLD_ERR_INT_RAW_W<'_, L1_CACHE_SYNC_PRELOAD_INT_RAW_SPEC>

Bit 10 - Reserved

pub fn l1_cache_pld_err_int_raw( &mut self ) -> L1_CACHE_PLD_ERR_INT_RAW_W<'_, L1_CACHE_SYNC_PRELOAD_INT_RAW_SPEC>

Bit 11 - The raw bit of the interrupt that occurs only when L1-Cache preload-operation error occurs.

pub fn cache_sync_err_int_raw( &mut self ) -> CACHE_SYNC_ERR_INT_RAW_W<'_, L1_CACHE_SYNC_PRELOAD_INT_RAW_SPEC>

Bit 13 - The raw bit of the interrupt that occurs only when Cache sync-operation error occurs.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual