Type Alias esp32c6::extmem::l2_cache_cachesize_conf::R

pub type R = R<L2_CACHE_CACHESIZE_CONF_SPEC>;

Register L2_CACHE_CACHESIZE_CONF reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn l2_cache_cachesize_1k(&self) -> L2_CACHE_CACHESIZE_1K_R

Bit 0 - The field is used to configure cachesize of L2-Cache as 1k bytes. This field and all other fields within this register is onehot.

pub fn l2_cache_cachesize_2k(&self) -> L2_CACHE_CACHESIZE_2K_R

Bit 1 - The field is used to configure cachesize of L2-Cache as 2k bytes. This field and all other fields within this register is onehot.

pub fn l2_cache_cachesize_4k(&self) -> L2_CACHE_CACHESIZE_4K_R

Bit 2 - The field is used to configure cachesize of L2-Cache as 4k bytes. This field and all other fields within this register is onehot.

pub fn l2_cache_cachesize_8k(&self) -> L2_CACHE_CACHESIZE_8K_R

Bit 3 - The field is used to configure cachesize of L2-Cache as 8k bytes. This field and all other fields within this register is onehot.

pub fn l2_cache_cachesize_16k(&self) -> L2_CACHE_CACHESIZE_16K_R

Bit 4 - The field is used to configure cachesize of L2-Cache as 16k bytes. This field and all other fields within this register is onehot.

pub fn l2_cache_cachesize_32k(&self) -> L2_CACHE_CACHESIZE_32K_R

Bit 5 - The field is used to configure cachesize of L2-Cache as 32k bytes. This field and all other fields within this register is onehot.

pub fn l2_cache_cachesize_64k(&self) -> L2_CACHE_CACHESIZE_64K_R

Bit 6 - The field is used to configure cachesize of L2-Cache as 64k bytes. This field and all other fields within this register is onehot.

pub fn l2_cache_cachesize_128k(&self) -> L2_CACHE_CACHESIZE_128K_R

Bit 7 - The field is used to configure cachesize of L2-Cache as 128k bytes. This field and all other fields within this register is onehot.

pub fn l2_cache_cachesize_256k(&self) -> L2_CACHE_CACHESIZE_256K_R

Bit 8 - The field is used to configure cachesize of L2-Cache as 256k bytes. This field and all other fields within this register is onehot.

pub fn l2_cache_cachesize_512k(&self) -> L2_CACHE_CACHESIZE_512K_R

Bit 9 - The field is used to configure cachesize of L2-Cache as 512k bytes. This field and all other fields within this register is onehot.

pub fn l2_cache_cachesize_1024k(&self) -> L2_CACHE_CACHESIZE_1024K_R

Bit 10 - The field is used to configure cachesize of L2-Cache as 1024k bytes. This field and all other fields within this register is onehot.

pub fn l2_cache_cachesize_2048k(&self) -> L2_CACHE_CACHESIZE_2048K_R

Bit 11 - The field is used to configure cachesize of L2-Cache as 2048k bytes. This field and all other fields within this register is onehot.

pub fn l2_cache_cachesize_4096k(&self) -> L2_CACHE_CACHESIZE_4096K_R

Bit 12 - The field is used to configure cachesize of L2-Cache as 4096k bytes. This field and all other fields within this register is onehot.