Module esp32c6::hinf

HINF Peripheral

Modules

cfg_data0

Configure sdio cis content

cfg_data1

SDIO configuration register

cfg_data7

SDIO configuration register

cfg_data16

SDIO cis configuration register

cfg_timing

Timing configuration registers

cfg_uhs1_int_mode

configure int to start and end ahead of time in uhs1 mode

cfg_update

update sdio configurations

cis_conf_w0

SDIO cis configuration register

cis_conf_w1

SDIO cis configuration register

cis_conf_w2

SDIO cis configuration register

cis_conf_w3

SDIO cis configuration register

cis_conf_w4

SDIO cis configuration register

cis_conf_w5

SDIO cis configuration register

cis_conf_w6

SDIO cis configuration register

cis_conf_w7

SDIO cis configuration register

conf_status

func0 config0 status

sdio_date

******* Description ***********

sdio_slave_eco_conf

sdio_slave redundant control registers

sdio_slave_eco_high

sdio_slave redundant control registers

sdio_slave_eco_low

sdio_slave redundant control registers

sdio_slave_ldo_conf

sdio slave ldo control register

Structs

RegisterBlock

Register block

Type Aliases

CFG_DATA0

CFG_DATA0 (rw) register accessor: Configure sdio cis content

CFG_DATA1

CFG_DATA1 (rw) register accessor: SDIO configuration register

CFG_DATA7

CFG_DATA7 (rw) register accessor: SDIO configuration register

CFG_DATA16

CFG_DATA16 (rw) register accessor: SDIO cis configuration register

CFG_TIMING

CFG_TIMING (rw) register accessor: Timing configuration registers

CFG_UHS1_INT_MODE

CFG_UHS1_INT_MODE (rw) register accessor: configure int to start and end ahead of time in uhs1 mode

CFG_UPDATE

CFG_UPDATE (w) register accessor: update sdio configurations

CIS_CONF_W0

CIS_CONF_W0 (rw) register accessor: SDIO cis configuration register

CIS_CONF_W1

CIS_CONF_W1 (rw) register accessor: SDIO cis configuration register

CIS_CONF_W2

CIS_CONF_W2 (rw) register accessor: SDIO cis configuration register

CIS_CONF_W3

CIS_CONF_W3 (rw) register accessor: SDIO cis configuration register

CIS_CONF_W4

CIS_CONF_W4 (rw) register accessor: SDIO cis configuration register

CIS_CONF_W5

CIS_CONF_W5 (rw) register accessor: SDIO cis configuration register

CIS_CONF_W6

CIS_CONF_W6 (rw) register accessor: SDIO cis configuration register

CIS_CONF_W7

CIS_CONF_W7 (rw) register accessor: SDIO cis configuration register

CONF_STATUS

CONF_STATUS (r) register accessor: func0 config0 status

SDIO_DATE

SDIO_DATE (rw) register accessor: ******* Description ***********

SDIO_SLAVE_ECO_CONF

SDIO_SLAVE_ECO_CONF (rw) register accessor: sdio_slave redundant control registers

SDIO_SLAVE_ECO_HIGH

SDIO_SLAVE_ECO_HIGH (rw) register accessor: sdio_slave redundant control registers

SDIO_SLAVE_ECO_LOW

SDIO_SLAVE_ECO_LOW (rw) register accessor: sdio_slave redundant control registers

SDIO_SLAVE_LDO_CONF

SDIO_SLAVE_LDO_CONF (rw) register accessor: sdio slave ldo control register