Type Alias esp32c6::ledc::conf::R

pub type R = R<CONF_SPEC>;

Register CONF reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn apb_clk_sel(&self) -> APB_CLK_SEL_R

Bits 0:1 - This bit is used to select clock source for the 4 timers . 2’d1: APB_CLK 2’d2: RTC8M_CLK 2’d3: XTAL_CLK

pub fn gamma_ram_clk_en_ch0(&self) -> GAMMA_RAM_CLK_EN_CH0_R

Bit 2 - This bit is used to control clock. 1’b1: Force clock on for gamma ram. 1’h0: Support clock only when application writes or read gamma ram.

pub fn gamma_ram_clk_en_ch1(&self) -> GAMMA_RAM_CLK_EN_CH1_R

Bit 3 - This bit is used to control clock. 1’b1: Force clock on for gamma ram. 1’h0: Support clock only when application writes or read gamma ram.

pub fn gamma_ram_clk_en_ch2(&self) -> GAMMA_RAM_CLK_EN_CH2_R

Bit 4 - This bit is used to control clock. 1’b1: Force clock on for gamma ram. 1’h0: Support clock only when application writes or read gamma ram.

pub fn gamma_ram_clk_en_ch3(&self) -> GAMMA_RAM_CLK_EN_CH3_R

Bit 5 - This bit is used to control clock. 1’b1: Force clock on for gamma ram. 1’h0: Support clock only when application writes or read gamma ram.

pub fn gamma_ram_clk_en_ch4(&self) -> GAMMA_RAM_CLK_EN_CH4_R

Bit 6 - This bit is used to control clock. 1’b1: Force clock on for gamma ram. 1’h0: Support clock only when application writes or read gamma ram.

pub fn gamma_ram_clk_en_ch5(&self) -> GAMMA_RAM_CLK_EN_CH5_R

Bit 7 - This bit is used to control clock. 1’b1: Force clock on for gamma ram. 1’h0: Support clock only when application writes or read gamma ram.

pub fn clk_en(&self) -> CLK_EN_R

Bit 31 - This bit is used to control clock. 1’b1: Force clock on for register. 1’h0: Support clock only when application writes registers.