Module esp32c6::mcpwm0

Motor Control Pulse-Width Modulation 0

Modules

cap_ch0

ch0 capture value status register

cap_ch0_cfg

Capture channel 0 configuration and enable

cap_ch1

ch1 capture value status register

cap_ch1_cfg

Capture channel 1 configuration and enable

cap_ch2

ch2 capture value status register

cap_ch2_cfg

Capture channel 2 configuration and enable

cap_status

Edge of last capture trigger

cap_timer_cfg

Configure capture timer

cap_timer_phase

Phase for capture timer sync

carrier0_cfg

Carrier enable and configuratoin

carrier1_cfg

Carrier enable and configuratoin

carrier2_cfg

Carrier enable and configuratoin

clk

MCPWM APB configuration register

clk_cfg

PWM clock prescaler register.

dt0_cfg

dead time type selection and configuration

dt0_fed_cfg

Shadow register for falling edge delay (FED).

dt0_red_cfg

Shadow register for rising edge delay (RED).

dt1_cfg

dead time type selection and configuration

dt1_fed_cfg

Shadow register for falling edge delay (FED).

dt1_red_cfg

Shadow register for rising edge delay (RED).

dt2_cfg

dead time type selection and configuration

dt2_fed_cfg

Shadow register for falling edge delay (FED).

dt2_red_cfg

Shadow register for rising edge delay (RED).

evt_en

MCPWM event enable register

fault_detect

Fault detection configuration and status

fh0_cfg0

Actions on PWM0A and PWM0B trip events

fh0_cfg1

Software triggers for fault handler actions

fh0_status

Status of fault events.

fh1_cfg0

Actions on PWM1A and PWM1B trip events

fh1_cfg1

Software triggers for fault handler actions

fh1_status

Status of fault events.

fh2_cfg0

Actions on PWM2A and PWM2B trip events

fh2_cfg1

Software triggers for fault handler actions

fh2_status

Status of fault events.

gen0_a

Actions triggered by events on PWM0A

gen0_b

Actions triggered by events on PWM0B

gen0_cfg0

Fault event T0 and T1 handling

gen0_force

Permissives to force PWM0A and PWM0B outputs by software

gen0_stmp_cfg

Transfer status and update method for time stamp registers A and B

gen0_tstmp_a

Shadow register for register A.

gen0_tstmp_b

Shadow register for register B.

gen1_a

Actions triggered by events on PWM1A

gen1_b

Actions triggered by events on PWM1B

gen1_cfg0

Fault event T0 and T1 handling

gen1_force

Permissives to force PWM1A and PWM1B outputs by software

gen1_stmp_cfg

Transfer status and update method for time stamp registers A and B

gen1_tstmp_a

Shadow register for register A.

gen1_tstmp_b

Shadow register for register B.

gen2_a

Actions triggered by events on PWM2A

gen2_b

Actions triggered by events on PWM2B

gen2_cfg0

Fault event T0 and T1 handling

gen2_force

Permissives to force PWM2A and PWM2B outputs by software

gen2_stmp_cfg

Transfer status and update method for time stamp registers A and B

gen2_tstmp_a

Shadow register for register A.

gen2_tstmp_b

Shadow register for register B.

int_clr

Interrupt clear bits

int_ena

Interrupt enable bits

int_raw

Raw interrupt status

int_st

Masked interrupt status

operator_timersel

Select specific timer for PWM operators.

task_en

MCPWM task enable register

timer0_cfg0

PWM timer0 period and update method configuration register.

timer0_cfg1

PWM timer0 working mode and start/stop control configuration register.

timer0_status

PWM timer0 status register.

timer0_sync

PWM timer0 sync function configuration register.

timer1_cfg0

PWM timer1 period and update method configuration register.

timer1_cfg1

PWM timer1 working mode and start/stop control configuration register.

timer1_status

PWM timer1 status register.

timer1_sync

PWM timer1 sync function configuration register.

timer2_cfg0

PWM timer2 period and update method configuration register.

timer2_cfg1

PWM timer2 working mode and start/stop control configuration register.

timer2_status

PWM timer2 status register.

timer2_sync

PWM timer2 sync function configuration register.

timer_synci_cfg

Synchronization input selection for three PWM timers.

update_cfg

Enable update.

version

Version register.

Structs

RegisterBlock

Register block

Type Aliases

CAP_CH0

CAP_CH0 (r) register accessor: ch0 capture value status register

CAP_CH0_CFG

CAP_CH0_CFG (rw) register accessor: Capture channel 0 configuration and enable

CAP_CH1

CAP_CH1 (r) register accessor: ch1 capture value status register

CAP_CH1_CFG

CAP_CH1_CFG (rw) register accessor: Capture channel 1 configuration and enable

CAP_CH2

CAP_CH2 (r) register accessor: ch2 capture value status register

CAP_CH2_CFG

CAP_CH2_CFG (rw) register accessor: Capture channel 2 configuration and enable

CAP_STATUS

CAP_STATUS (r) register accessor: Edge of last capture trigger

CAP_TIMER_CFG

CAP_TIMER_CFG (rw) register accessor: Configure capture timer

CAP_TIMER_PHASE

CAP_TIMER_PHASE (rw) register accessor: Phase for capture timer sync

CARRIER0_CFG

CARRIER0_CFG (rw) register accessor: Carrier enable and configuratoin

CARRIER1_CFG

CARRIER1_CFG (rw) register accessor: Carrier enable and configuratoin

CARRIER2_CFG

CARRIER2_CFG (rw) register accessor: Carrier enable and configuratoin

CLK

CLK (rw) register accessor: MCPWM APB configuration register

CLK_CFG

CLK_CFG (rw) register accessor: PWM clock prescaler register.

DT0_CFG

DT0_CFG (rw) register accessor: dead time type selection and configuration

DT0_FED_CFG

DT0_FED_CFG (rw) register accessor: Shadow register for falling edge delay (FED).

DT0_RED_CFG

DT0_RED_CFG (rw) register accessor: Shadow register for rising edge delay (RED).

DT1_CFG

DT1_CFG (rw) register accessor: dead time type selection and configuration

DT1_FED_CFG

DT1_FED_CFG (rw) register accessor: Shadow register for falling edge delay (FED).

DT1_RED_CFG

DT1_RED_CFG (rw) register accessor: Shadow register for rising edge delay (RED).

DT2_CFG

DT2_CFG (rw) register accessor: dead time type selection and configuration

DT2_FED_CFG

DT2_FED_CFG (rw) register accessor: Shadow register for falling edge delay (FED).

DT2_RED_CFG

DT2_RED_CFG (rw) register accessor: Shadow register for rising edge delay (RED).

EVT_EN

EVT_EN (rw) register accessor: MCPWM event enable register

FAULT_DETECT

FAULT_DETECT (rw) register accessor: Fault detection configuration and status

FH0_CFG0

FH0_CFG0 (rw) register accessor: Actions on PWM0A and PWM0B trip events

FH0_CFG1

FH0_CFG1 (rw) register accessor: Software triggers for fault handler actions

FH0_STATUS

FH0_STATUS (r) register accessor: Status of fault events.

FH1_CFG0

FH1_CFG0 (rw) register accessor: Actions on PWM1A and PWM1B trip events

FH1_CFG1

FH1_CFG1 (rw) register accessor: Software triggers for fault handler actions

FH1_STATUS

FH1_STATUS (r) register accessor: Status of fault events.

FH2_CFG0

FH2_CFG0 (rw) register accessor: Actions on PWM2A and PWM2B trip events

FH2_CFG1

FH2_CFG1 (rw) register accessor: Software triggers for fault handler actions

FH2_STATUS

FH2_STATUS (r) register accessor: Status of fault events.

GEN0_A

GEN0_A (rw) register accessor: Actions triggered by events on PWM0A

GEN0_B

GEN0_B (rw) register accessor: Actions triggered by events on PWM0B

GEN0_CFG0

GEN0_CFG0 (rw) register accessor: Fault event T0 and T1 handling

GEN0_FORCE

GEN0_FORCE (rw) register accessor: Permissives to force PWM0A and PWM0B outputs by software

GEN0_STMP_CFG

GEN0_STMP_CFG (rw) register accessor: Transfer status and update method for time stamp registers A and B

GEN0_TSTMP_A

GEN0_TSTMP_A (rw) register accessor: Shadow register for register A.

GEN0_TSTMP_B

GEN0_TSTMP_B (rw) register accessor: Shadow register for register B.

GEN1_A

GEN1_A (rw) register accessor: Actions triggered by events on PWM1A

GEN1_B

GEN1_B (rw) register accessor: Actions triggered by events on PWM1B

GEN1_CFG0

GEN1_CFG0 (rw) register accessor: Fault event T0 and T1 handling

GEN1_FORCE

GEN1_FORCE (rw) register accessor: Permissives to force PWM1A and PWM1B outputs by software

GEN1_STMP_CFG

GEN1_STMP_CFG (rw) register accessor: Transfer status and update method for time stamp registers A and B

GEN1_TSTMP_A

GEN1_TSTMP_A (rw) register accessor: Shadow register for register A.

GEN1_TSTMP_B

GEN1_TSTMP_B (rw) register accessor: Shadow register for register B.

GEN2_A

GEN2_A (rw) register accessor: Actions triggered by events on PWM2A

GEN2_B

GEN2_B (rw) register accessor: Actions triggered by events on PWM2B

GEN2_CFG0

GEN2_CFG0 (rw) register accessor: Fault event T0 and T1 handling

GEN2_FORCE

GEN2_FORCE (rw) register accessor: Permissives to force PWM2A and PWM2B outputs by software

GEN2_STMP_CFG

GEN2_STMP_CFG (rw) register accessor: Transfer status and update method for time stamp registers A and B

GEN2_TSTMP_A

GEN2_TSTMP_A (rw) register accessor: Shadow register for register A.

GEN2_TSTMP_B

GEN2_TSTMP_B (rw) register accessor: Shadow register for register B.

INT_CLR

INT_CLR (w) register accessor: Interrupt clear bits

INT_ENA

INT_ENA (rw) register accessor: Interrupt enable bits

INT_RAW

INT_RAW (rw) register accessor: Raw interrupt status

INT_ST

INT_ST (r) register accessor: Masked interrupt status

OPERATOR_TIMERSEL

OPERATOR_TIMERSEL (rw) register accessor: Select specific timer for PWM operators.

TASK_EN

TASK_EN (rw) register accessor: MCPWM task enable register

TIMER0_CFG0

TIMER0_CFG0 (rw) register accessor: PWM timer0 period and update method configuration register.

TIMER0_CFG1

TIMER0_CFG1 (rw) register accessor: PWM timer0 working mode and start/stop control configuration register.

TIMER0_STATUS

TIMER0_STATUS (r) register accessor: PWM timer0 status register.

TIMER0_SYNC

TIMER0_SYNC (rw) register accessor: PWM timer0 sync function configuration register.

TIMER1_CFG0

TIMER1_CFG0 (rw) register accessor: PWM timer1 period and update method configuration register.

TIMER1_CFG1

TIMER1_CFG1 (rw) register accessor: PWM timer1 working mode and start/stop control configuration register.

TIMER1_STATUS

TIMER1_STATUS (r) register accessor: PWM timer1 status register.

TIMER1_SYNC

TIMER1_SYNC (rw) register accessor: PWM timer1 sync function configuration register.

TIMER2_CFG0

TIMER2_CFG0 (rw) register accessor: PWM timer2 period and update method configuration register.

TIMER2_CFG1

TIMER2_CFG1 (rw) register accessor: PWM timer2 working mode and start/stop control configuration register.

TIMER2_STATUS

TIMER2_STATUS (r) register accessor: PWM timer2 status register.

TIMER2_SYNC

TIMER2_SYNC (rw) register accessor: PWM timer2 sync function configuration register.

TIMER_SYNCI_CFG

TIMER_SYNCI_CFG (rw) register accessor: Synchronization input selection for three PWM timers.

UPDATE_CFG

UPDATE_CFG (rw) register accessor: Enable update.

VERSION

VERSION (rw) register accessor: Version register.