Module esp32c6::mcpwm0::fh1_cfg0

Actions on PWM1A and PWM1B trip events

Structs

FH1_CFG0_SPEC

Actions on PWM1A and PWM1B trip events

Type Aliases

R

Register FH1_CFG0 reader

TZ1_A_CBC_D_R

Field TZ1_A_CBC_D reader - Cycle-by-cycle mode action on PWM1A when fault event occurs and timer is decreasing. 0: do nothing, 1: force low, 2: force high, 3: toggle

TZ1_A_CBC_D_W

Field TZ1_A_CBC_D writer - Cycle-by-cycle mode action on PWM1A when fault event occurs and timer is decreasing. 0: do nothing, 1: force low, 2: force high, 3: toggle

TZ1_A_CBC_U_R

Field TZ1_A_CBC_U reader - Cycle-by-cycle mode action on PWM1A when fault event occurs and timer is increasing. 0: do nothing, 1: force low, 2: force high, 3: toggle

TZ1_A_CBC_U_W

Field TZ1_A_CBC_U writer - Cycle-by-cycle mode action on PWM1A when fault event occurs and timer is increasing. 0: do nothing, 1: force low, 2: force high, 3: toggle

TZ1_A_OST_D_R

Field TZ1_A_OST_D reader - One-shot mode action on PWM1A when fault event occurs and timer is decreasing. 0: do nothing,1: force low, 2: force high, 3: toggle

TZ1_A_OST_D_W

Field TZ1_A_OST_D writer - One-shot mode action on PWM1A when fault event occurs and timer is decreasing. 0: do nothing,1: force low, 2: force high, 3: toggle

TZ1_A_OST_U_R

Field TZ1_A_OST_U reader - One-shot mode action on PWM1A when fault event occurs and timer is increasing. 0: do nothing, 1: force low, 2: force high, 3: toggle

TZ1_A_OST_U_W

Field TZ1_A_OST_U writer - One-shot mode action on PWM1A when fault event occurs and timer is increasing. 0: do nothing, 1: force low, 2: force high, 3: toggle

TZ1_B_CBC_D_R

Field TZ1_B_CBC_D reader - Cycle-by-cycle mode action on PWM1B when fault event occurs and timer is decreasing. 0: do nothing, 1: force low, 2: force high, 3: toggle

TZ1_B_CBC_D_W

Field TZ1_B_CBC_D writer - Cycle-by-cycle mode action on PWM1B when fault event occurs and timer is decreasing. 0: do nothing, 1: force low, 2: force high, 3: toggle

TZ1_B_CBC_U_R

Field TZ1_B_CBC_U reader - Cycle-by-cycle mode action on PWM1B when fault event occurs and timer is increasing. 0: do nothing, 1: force low, 2: force high, 3: toggle

TZ1_B_CBC_U_W

Field TZ1_B_CBC_U writer - Cycle-by-cycle mode action on PWM1B when fault event occurs and timer is increasing. 0: do nothing, 1: force low, 2: force high, 3: toggle

TZ1_B_OST_D_R

Field TZ1_B_OST_D reader - One-shot mode action on PWM1B when fault event occurs and timer is decreasing. 0: do nothing, 1: force low, 2: force high, 3: toggle

TZ1_B_OST_D_W

Field TZ1_B_OST_D writer - One-shot mode action on PWM1B when fault event occurs and timer is decreasing. 0: do nothing, 1: force low, 2: force high, 3: toggle

TZ1_B_OST_U_R

Field TZ1_B_OST_U reader - One-shot mode action on PWM1B when fault event occurs and timer is increasing. 0: do nothing, 1: force low, 2: force high, 3: toggle

TZ1_B_OST_U_W

Field TZ1_B_OST_U writer - One-shot mode action on PWM1B when fault event occurs and timer is increasing. 0: do nothing, 1: force low, 2: force high, 3: toggle

TZ1_F0_CBC_R

Field TZ1_F0_CBC reader - event_f0 will trigger cycle-by-cycle mode action. 0: disable, 1: enable

TZ1_F0_CBC_W

Field TZ1_F0_CBC writer - event_f0 will trigger cycle-by-cycle mode action. 0: disable, 1: enable

TZ1_F0_OST_R

Field TZ1_F0_OST reader - event_f0 will trigger one-shot mode action. 0: disable, 1: enable

TZ1_F0_OST_W

Field TZ1_F0_OST writer - event_f0 will trigger one-shot mode action. 0: disable, 1: enable

TZ1_F1_CBC_R

Field TZ1_F1_CBC reader - event_f1 will trigger cycle-by-cycle mode action. 0: disable, 1: enable

TZ1_F1_CBC_W

Field TZ1_F1_CBC writer - event_f1 will trigger cycle-by-cycle mode action. 0: disable, 1: enable

TZ1_F1_OST_R

Field TZ1_F1_OST reader - event_f1 will trigger one-shot mode action. 0: disable, 1: enable

TZ1_F1_OST_W

Field TZ1_F1_OST writer - event_f1 will trigger one-shot mode action. 0: disable, 1: enable

TZ1_F2_CBC_R

Field TZ1_F2_CBC reader - event_f2 will trigger cycle-by-cycle mode action. 0: disable, 1: enable

TZ1_F2_CBC_W

Field TZ1_F2_CBC writer - event_f2 will trigger cycle-by-cycle mode action. 0: disable, 1: enable

TZ1_F2_OST_R

Field TZ1_F2_OST reader - event_f2 will trigger one-shot mode action. 0: disable, 1: enable

TZ1_F2_OST_W

Field TZ1_F2_OST writer - event_f2 will trigger one-shot mode action. 0: disable, 1: enable

TZ1_SW_CBC_R

Field TZ1_SW_CBC reader - Enable register for software force cycle-by-cycle mode action. 0: disable, 1: enable

TZ1_SW_CBC_W

Field TZ1_SW_CBC writer - Enable register for software force cycle-by-cycle mode action. 0: disable, 1: enable

TZ1_SW_OST_R

Field TZ1_SW_OST reader - Enable register for software force one-shot mode action. 0: disable, 1: enable

TZ1_SW_OST_W

Field TZ1_SW_OST writer - Enable register for software force one-shot mode action. 0: disable, 1: enable

W

Register FH1_CFG0 writer