Type Alias esp32c6::mcpwm0::fh1_cfg0::R

pub type R = R<FH1_CFG0_SPEC>;

Register FH1_CFG0 reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn tz1_sw_cbc(&self) -> TZ1_SW_CBC_R

Bit 0 - Enable register for software force cycle-by-cycle mode action. 0: disable, 1: enable

pub fn tz1_f2_cbc(&self) -> TZ1_F2_CBC_R

Bit 1 - event_f2 will trigger cycle-by-cycle mode action. 0: disable, 1: enable

pub fn tz1_f1_cbc(&self) -> TZ1_F1_CBC_R

Bit 2 - event_f1 will trigger cycle-by-cycle mode action. 0: disable, 1: enable

pub fn tz1_f0_cbc(&self) -> TZ1_F0_CBC_R

Bit 3 - event_f0 will trigger cycle-by-cycle mode action. 0: disable, 1: enable

pub fn tz1_sw_ost(&self) -> TZ1_SW_OST_R

Bit 4 - Enable register for software force one-shot mode action. 0: disable, 1: enable

pub fn tz1_f2_ost(&self) -> TZ1_F2_OST_R

Bit 5 - event_f2 will trigger one-shot mode action. 0: disable, 1: enable

pub fn tz1_f1_ost(&self) -> TZ1_F1_OST_R

Bit 6 - event_f1 will trigger one-shot mode action. 0: disable, 1: enable

pub fn tz1_f0_ost(&self) -> TZ1_F0_OST_R

Bit 7 - event_f0 will trigger one-shot mode action. 0: disable, 1: enable

pub fn tz1_a_cbc_d(&self) -> TZ1_A_CBC_D_R

Bits 8:9 - Cycle-by-cycle mode action on PWM1A when fault event occurs and timer is decreasing. 0: do nothing, 1: force low, 2: force high, 3: toggle

pub fn tz1_a_cbc_u(&self) -> TZ1_A_CBC_U_R

Bits 10:11 - Cycle-by-cycle mode action on PWM1A when fault event occurs and timer is increasing. 0: do nothing, 1: force low, 2: force high, 3: toggle

pub fn tz1_a_ost_d(&self) -> TZ1_A_OST_D_R

Bits 12:13 - One-shot mode action on PWM1A when fault event occurs and timer is decreasing. 0: do nothing,1: force low, 2: force high, 3: toggle

pub fn tz1_a_ost_u(&self) -> TZ1_A_OST_U_R

Bits 14:15 - One-shot mode action on PWM1A when fault event occurs and timer is increasing. 0: do nothing, 1: force low, 2: force high, 3: toggle

pub fn tz1_b_cbc_d(&self) -> TZ1_B_CBC_D_R

Bits 16:17 - Cycle-by-cycle mode action on PWM1B when fault event occurs and timer is decreasing. 0: do nothing, 1: force low, 2: force high, 3: toggle

pub fn tz1_b_cbc_u(&self) -> TZ1_B_CBC_U_R

Bits 18:19 - Cycle-by-cycle mode action on PWM1B when fault event occurs and timer is increasing. 0: do nothing, 1: force low, 2: force high, 3: toggle

pub fn tz1_b_ost_d(&self) -> TZ1_B_OST_D_R

Bits 20:21 - One-shot mode action on PWM1B when fault event occurs and timer is decreasing. 0: do nothing, 1: force low, 2: force high, 3: toggle

pub fn tz1_b_ost_u(&self) -> TZ1_B_OST_U_R

Bits 22:23 - One-shot mode action on PWM1B when fault event occurs and timer is increasing. 0: do nothing, 1: force low, 2: force high, 3: toggle