Type Alias esp32c6::mcpwm0::update_cfg::W

pub type W = W<UPDATE_CFG_SPEC>;

Register UPDATE_CFG writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn global_up_en(&mut self) -> GLOBAL_UP_EN_W<'_, UPDATE_CFG_SPEC>

Bit 0 - The global enable of update of all active registers in MCPWM module

pub fn global_force_up(&mut self) -> GLOBAL_FORCE_UP_W<'_, UPDATE_CFG_SPEC>

Bit 1 - a toggle (software invert its value) will trigger a forced update of all active registers in MCPWM module

pub fn op0_up_en(&mut self) -> OP0_UP_EN_W<'_, UPDATE_CFG_SPEC>

Bit 2 - When set and PWM_GLOBAL_UP_EN is set, update of active registers in PWM operator 0 are enabled

pub fn op0_force_up(&mut self) -> OP0_FORCE_UP_W<'_, UPDATE_CFG_SPEC>

Bit 3 - a toggle (software invert its value) will trigger a forced update of active registers in PWM operator 0

pub fn op1_up_en(&mut self) -> OP1_UP_EN_W<'_, UPDATE_CFG_SPEC>

Bit 4 - When set and PWM_GLOBAL_UP_EN is set, update of active registers in PWM operator 1 are enabled

pub fn op1_force_up(&mut self) -> OP1_FORCE_UP_W<'_, UPDATE_CFG_SPEC>

Bit 5 - a toggle (software invert its value) will trigger a forced update of active registers in PWM operator 1

pub fn op2_up_en(&mut self) -> OP2_UP_EN_W<'_, UPDATE_CFG_SPEC>

Bit 6 - When set and PWM_GLOBAL_UP_EN is set, update of active registers in PWM operator 2 are enabled

pub fn op2_force_up(&mut self) -> OP2_FORCE_UP_W<'_, UPDATE_CFG_SPEC>

Bit 7 - a toggle (software invert its value) will trigger a forced update of active registers in PWM operator 2

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual