Module esp32c6::parl_io::rx_cfg0

Parallel RX module configuration register0.

Structs

RX_CFG0_SPEC

Parallel RX module configuration register0.

Type Aliases

R

Register RX_CFG0 reader

RX_BIT_PACK_ORDER_R

Field RX_BIT_PACK_ORDER reader - Write 0 to pack bits into 1byte from MSB when data bus width is 4/2/1 bits. Write 0 to pack bits into 1byte from LSB when data bus width is 4/2/1 bits.

RX_BIT_PACK_ORDER_W

Field RX_BIT_PACK_ORDER writer - Write 0 to pack bits into 1byte from MSB when data bus width is 4/2/1 bits. Write 0 to pack bits into 1byte from LSB when data bus width is 4/2/1 bits.

RX_BUS_WID_SEL_R

Field RX_BUS_WID_SEL reader - Rx data bus width selection. 100: bus width is 1 bit 011: bus width is 2 bits 010: bus width is 4 bits 001: bus width is 8 bits 000: bus width is 16 bits

RX_BUS_WID_SEL_W

Field RX_BUS_WID_SEL writer - Rx data bus width selection. 100: bus width is 1 bit 011: bus width is 2 bits 010: bus width is 4 bits 001: bus width is 8 bits 000: bus width is 16 bits

RX_CLK_EDGE_SEL_R

Field RX_CLK_EDGE_SEL reader - Write 0 to enable sampling data on the rising edge of rx clock. Write 0 to enable sampling data on the falling edge of rx clock.

RX_CLK_EDGE_SEL_W

Field RX_CLK_EDGE_SEL writer - Write 0 to enable sampling data on the rising edge of rx clock. Write 0 to enable sampling data on the falling edge of rx clock.

RX_DATA_BYTELEN_R

Field RX_DATA_BYTELEN reader - Configures rx receieved data byte length.

RX_DATA_BYTELEN_W

Field RX_DATA_BYTELEN writer - Configures rx receieved data byte length.

RX_EOF_GEN_SEL_R

Field RX_EOF_GEN_SEL reader - Write 0 to select eof generated manchnism by configured data byte length. Write 1 to select eof generated manchnism by external enable signal.

RX_EOF_GEN_SEL_W

Field RX_EOF_GEN_SEL writer - Write 0 to select eof generated manchnism by configured data byte length. Write 1 to select eof generated manchnism by external enable signal.

RX_FIFO_SRST_R

Field RX_FIFO_SRST reader - Write 1 to enable soft reset of async fifo in rx module.

RX_FIFO_SRST_W

Field RX_FIFO_SRST writer - Write 1 to enable soft reset of async fifo in rx module.

RX_LEVEL_SUBMODE_SEL_R

Field RX_LEVEL_SUBMODE_SEL reader - Write 0 to sample data at high level of external enable signal. Write 1 to sample data at low level of external enable signal.

RX_LEVEL_SUBMODE_SEL_W

Field RX_LEVEL_SUBMODE_SEL writer - Write 0 to sample data at high level of external enable signal. Write 1 to sample data at low level of external enable signal.

RX_PULSE_SUBMODE_SEL_R

Field RX_PULSE_SUBMODE_SEL reader - Pulse submode selection. 0000: positive pulse start(data bit included) && positive pulse end(data bit included) 0001: positive pulse start(data bit included) && positive pulse end (data bit excluded) 0010: positive pulse start(data bit excluded) && positive pulse end (data bit included) 0011: positive pulse start(data bit excluded) && positive pulse end (data bit excluded) 0100: positive pulse start(data bit included) && length end 0101: positive pulse start(data bit excluded) && length end 0110: negative pulse start(data bit included) && negative pulse end(data bit included) 0111: negative pulse start(data bit included) && negative pulse end (data bit excluded) 1000: negative pulse start(data bit excluded) && negative pulse end (data bit included) 1001: negative pulse start(data bit excluded) && negative pulse end (data bit excluded) 1010: negative pulse start(data bit included) && length end 1011: negative pulse start(data bit excluded) && length end

RX_PULSE_SUBMODE_SEL_W

Field RX_PULSE_SUBMODE_SEL writer - Pulse submode selection. 0000: positive pulse start(data bit included) && positive pulse end(data bit included) 0001: positive pulse start(data bit included) && positive pulse end (data bit excluded) 0010: positive pulse start(data bit excluded) && positive pulse end (data bit included) 0011: positive pulse start(data bit excluded) && positive pulse end (data bit excluded) 0100: positive pulse start(data bit included) && length end 0101: positive pulse start(data bit excluded) && length end 0110: negative pulse start(data bit included) && negative pulse end(data bit included) 0111: negative pulse start(data bit included) && negative pulse end (data bit excluded) 1000: negative pulse start(data bit excluded) && negative pulse end (data bit included) 1001: negative pulse start(data bit excluded) && negative pulse end (data bit excluded) 1010: negative pulse start(data bit included) && length end 1011: negative pulse start(data bit excluded) && length end

RX_SMP_MODE_SEL_R

Field RX_SMP_MODE_SEL reader - Rx data sampling mode selection. 000: external level enable mode 001: external pulse enable mode 010: internal software enable mode

RX_SMP_MODE_SEL_W

Field RX_SMP_MODE_SEL writer - Rx data sampling mode selection. 000: external level enable mode 001: external pulse enable mode 010: internal software enable mode

RX_START_R

Field RX_START reader - Write 1 to start rx global data sampling.

RX_START_W

Field RX_START writer - Write 1 to start rx global data sampling.

RX_SW_EN_R

Field RX_SW_EN reader - Write 1 to enable software data sampling.

RX_SW_EN_W

Field RX_SW_EN writer - Write 1 to enable software data sampling.

W

Register RX_CFG0 writer