Type Alias esp32c6::parl_io::rx_cfg0::R

pub type R = R<RX_CFG0_SPEC>;

Register RX_CFG0 reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn rx_eof_gen_sel(&self) -> RX_EOF_GEN_SEL_R

Bit 0 - Write 0 to select eof generated manchnism by configured data byte length. Write 1 to select eof generated manchnism by external enable signal.

pub fn rx_start(&self) -> RX_START_R

Bit 1 - Write 1 to start rx global data sampling.

pub fn rx_data_bytelen(&self) -> RX_DATA_BYTELEN_R

Bits 2:17 - Configures rx receieved data byte length.

pub fn rx_sw_en(&self) -> RX_SW_EN_R

Bit 18 - Write 1 to enable software data sampling.

pub fn rx_pulse_submode_sel(&self) -> RX_PULSE_SUBMODE_SEL_R

Bits 19:22 - Pulse submode selection. 0000: positive pulse start(data bit included) && positive pulse end(data bit included) 0001: positive pulse start(data bit included) && positive pulse end (data bit excluded) 0010: positive pulse start(data bit excluded) && positive pulse end (data bit included) 0011: positive pulse start(data bit excluded) && positive pulse end (data bit excluded) 0100: positive pulse start(data bit included) && length end 0101: positive pulse start(data bit excluded) && length end 0110: negative pulse start(data bit included) && negative pulse end(data bit included) 0111: negative pulse start(data bit included) && negative pulse end (data bit excluded) 1000: negative pulse start(data bit excluded) && negative pulse end (data bit included) 1001: negative pulse start(data bit excluded) && negative pulse end (data bit excluded) 1010: negative pulse start(data bit included) && length end 1011: negative pulse start(data bit excluded) && length end

pub fn rx_level_submode_sel(&self) -> RX_LEVEL_SUBMODE_SEL_R

Bit 23 - Write 0 to sample data at high level of external enable signal. Write 1 to sample data at low level of external enable signal.

pub fn rx_smp_mode_sel(&self) -> RX_SMP_MODE_SEL_R

Bits 24:25 - Rx data sampling mode selection. 000: external level enable mode 001: external pulse enable mode 010: internal software enable mode

pub fn rx_clk_edge_sel(&self) -> RX_CLK_EDGE_SEL_R

Bit 26 - Write 0 to enable sampling data on the rising edge of rx clock. Write 0 to enable sampling data on the falling edge of rx clock.

pub fn rx_bit_pack_order(&self) -> RX_BIT_PACK_ORDER_R

Bit 27 - Write 0 to pack bits into 1byte from MSB when data bus width is 4/2/1 bits. Write 0 to pack bits into 1byte from LSB when data bus width is 4/2/1 bits.

pub fn rx_bus_wid_sel(&self) -> RX_BUS_WID_SEL_R

Bits 28:30 - Rx data bus width selection. 100: bus width is 1 bit 011: bus width is 2 bits 010: bus width is 4 bits 001: bus width is 8 bits 000: bus width is 16 bits

pub fn rx_fifo_srst(&self) -> RX_FIFO_SRST_R

Bit 31 - Write 1 to enable soft reset of async fifo in rx module.