Type Alias esp32c6::parl_io::tx_cfg0::R

pub type R = R<TX_CFG0_SPEC>;

Register TX_CFG0 reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn tx_bytelen(&self) -> TX_BYTELEN_R

Bits 2:17 - Configures tx sending data byte length.

pub fn tx_gating_en(&self) -> TX_GATING_EN_R

Bit 18 - Write 1 to enable output tx clock gating.

pub fn tx_start(&self) -> TX_START_R

Bit 19 - Write 1 to start tx global data output.

pub fn tx_hw_valid_en(&self) -> TX_HW_VALID_EN_R

Bit 20 - Write 1 to enable tx hardware data valid signal.

pub fn tx_smp_edge_sel(&self) -> TX_SMP_EDGE_SEL_R

Bit 25 - Write 0 to enable sampling data on the rising edge of tx clock. Write 0 to enable sampling data on the falling edge of tx clock.

pub fn tx_bit_unpack_order(&self) -> TX_BIT_UNPACK_ORDER_R

Bit 26 - Write 0 to unpack bits from 1byte from MSB when data bus width is 4/2/1 bits. Write 0 to unpack bits from 1byte from LSB when data bus width is 4/2/1 bits.

pub fn tx_bus_wid_sel(&self) -> TX_BUS_WID_SEL_R

Bits 27:29 - Tx data bus width selection. 100: bus width is 1 bit 011: bus width is 2 bits 010: bus width is 4 bits 001: bus width is 8 bits 000: bus width is 16 bits

pub fn tx_fifo_srst(&self) -> TX_FIFO_SRST_R

Bit 30 - Write 1 to enable soft reset of async fifo in tx module.