Struct esp32c6::sha::RegisterBlock

#[repr(C)]
pub struct RegisterBlock { /* private fields */ }

Register block

Implementations

impl RegisterBlock

pub const fn mode(&self) -> &MODE

0x00 - Initial configuration register.

pub const fn t_string(&self) -> &T_STRING

0x04 - SHA 512/t configuration register 0.

pub const fn t_length(&self) -> &T_LENGTH

0x08 - SHA 512/t configuration register 1.

pub const fn dma_block_num(&self) -> &DMA_BLOCK_NUM

0x0c - DMA configuration register 0.

pub const fn start(&self) -> &START

0x10 - Typical SHA configuration register 0.

pub const fn continue_(&self) -> &CONTINUE

0x14 - Typical SHA configuration register 1.

pub const fn busy(&self) -> &BUSY

0x18 - Busy register.

pub const fn dma_start(&self) -> &DMA_START

0x1c - DMA configuration register 1.

pub const fn dma_continue(&self) -> &DMA_CONTINUE

0x20 - DMA configuration register 2.

pub const fn clear_irq(&self) -> &CLEAR_IRQ

0x24 - Interrupt clear register.

pub const fn irq_ena(&self) -> &IRQ_ENA

0x28 - Interrupt enable register.

pub const fn date(&self) -> &DATE

0x2c - Date register.

pub const fn h_mem(&self, n: usize) -> &H_MEM

0x40..0x60 - Sha H memory which contains intermediate hash or finial hash.

pub fn h_mem_iter(&self) -> impl Iterator<Item = &H_MEM>

Iterator for array of: 0x40..0x60 - Sha H memory which contains intermediate hash or finial hash.

pub const fn m_mem(&self, n: usize) -> &M_MEM

0x80..0xc0 - Sha M memory which contains message.

pub fn m_mem_iter(&self) -> impl Iterator<Item = &M_MEM>

Iterator for array of: 0x80..0xc0 - Sha M memory which contains message.