Module esp32c6::spi0::spi_mem_sram_cmd

SPI0 external RAM mode control register

Structs

SPI_MEM_SRAM_CMD_SPEC

SPI0 external RAM mode control register

Type Aliases

R

Register SPI_MEM_SRAM_CMD reader

SPI_MEM_SADDR_DUAL_R

Field SPI_MEM_SADDR_DUAL reader - For SPI0 external RAM , address phase apply 2 signals. 1: enable 0: disable. The bit is the same with spi_mem_usr_sram_dio.

SPI_MEM_SADDR_OCT_R

Field SPI_MEM_SADDR_OCT reader - For SPI0 external RAM , address phase apply 4 signals. 1: enable 0: disable.

SPI_MEM_SADDR_QUAD_R

Field SPI_MEM_SADDR_QUAD reader - For SPI0 external RAM , address phase apply 4 signals. 1: enable 0: disable. The bit is the same with spi_mem_usr_sram_qio.

SPI_MEM_SCLK_MODE_R

Field SPI_MEM_SCLK_MODE reader - SPI clock mode bits. 0: SPI clock is off when CS inactive 1: SPI clock is delayed one cycle after CS inactive 2: SPI clock is delayed two cycles after CS inactive 3: SPI clock is always on.

SPI_MEM_SCMD_OCT_R

Field SPI_MEM_SCMD_OCT reader - For SPI0 external RAM , cmd phase apply 8 signals. 1: enable 0: disable.

SPI_MEM_SCMD_QUAD_R

Field SPI_MEM_SCMD_QUAD reader - For SPI0 external RAM , cmd phase apply 4 signals. 1: enable 0: disable. The bit is the same with spi_mem_usr_sram_qio.

SPI_MEM_SDIN_DUAL_R

Field SPI_MEM_SDIN_DUAL reader - For SPI0 external RAM , din phase apply 2 signals. 1: enable 0: disable. The bit is the same with spi_mem_usr_sram_dio.

SPI_MEM_SDIN_OCT_R

Field SPI_MEM_SDIN_OCT reader - For SPI0 external RAM , din phase apply 8 signals. 1: enable 0: disable.

SPI_MEM_SDIN_QUAD_R

Field SPI_MEM_SDIN_QUAD reader - For SPI0 external RAM , din phase apply 4 signals. 1: enable 0: disable. The bit is the same with spi_mem_usr_sram_qio.

SPI_MEM_SDOUT_DUAL_R

Field SPI_MEM_SDOUT_DUAL reader - For SPI0 external RAM , dout phase apply 2 signals. 1: enable 0: disable. The bit is the same with spi_mem_usr_sram_dio.

SPI_MEM_SDOUT_OCT_R

Field SPI_MEM_SDOUT_OCT reader - For SPI0 external RAM , dout phase apply 8 signals. 1: enable 0: disable.

SPI_MEM_SDOUT_QUAD_R

Field SPI_MEM_SDOUT_QUAD reader - For SPI0 external RAM , dout phase apply 4 signals. 1: enable 0: disable. The bit is the same with spi_mem_usr_sram_qio.

SPI_MEM_SDUMMY_RIN_R

Field SPI_MEM_SDUMMY_RIN reader - In the dummy phase of a MSPI read data transfer when accesses to external RAM, the signal level of SPI bus is output by the MSPI controller.

SPI_MEM_SDUMMY_RIN_W

Field SPI_MEM_SDUMMY_RIN writer - In the dummy phase of a MSPI read data transfer when accesses to external RAM, the signal level of SPI bus is output by the MSPI controller.

SPI_MEM_SDUMMY_WOUT_R

Field SPI_MEM_SDUMMY_WOUT reader - In the dummy phase of a MSPI write data transfer when accesses to external RAM, the signal level of SPI bus is output by the MSPI controller.

SPI_MEM_SWB_MODE_R

Field SPI_MEM_SWB_MODE reader - Mode bits in the external RAM fast read mode it is combined with spi_mem_fastrd_mode bit.

SPI_SMEM_DATA_IE_ALWAYS_ON_R

Field SPI_SMEM_DATA_IE_ALWAYS_ON reader - When accesses to external RAM, 1: the IE signals of pads connected to SPI_IO[7:0] are always 1. 0: Others.

SPI_SMEM_DQS_IE_ALWAYS_ON_R

Field SPI_SMEM_DQS_IE_ALWAYS_ON reader - When accesses to external RAM, 1: the IE signals of pads connected to SPI_DQS are always 1. 0: Others.

SPI_SMEM_WDUMMY_ALWAYS_OUT_R

Field SPI_SMEM_WDUMMY_ALWAYS_OUT reader - In the dummy phase of an MSPI write data transfer when accesses to external RAM, the level of SPI_IO[7:0] is output by the MSPI controller.

SPI_SMEM_WDUMMY_DQS_ALWAYS_OUT_R

Field SPI_SMEM_WDUMMY_DQS_ALWAYS_OUT reader - In the dummy phase of an MSPI write data transfer when accesses to external RAM, the level of SPI_DQS is output by the MSPI controller.

W

Register SPI_MEM_SRAM_CMD writer