Module esp32c6::spi0::spi_smem_ac

MSPI external RAM ECC and SPI CS timing control register

Structs

SPI_SMEM_AC_SPEC

MSPI external RAM ECC and SPI CS timing control register

Type Aliases

R

Register SPI_SMEM_AC reader

SPI_SMEM_CS_HOLD_DELAY_R

Field SPI_SMEM_CS_HOLD_DELAY reader - These bits are used to set the minimum CS high time tSHSL between SPI burst transfer when accesses to external RAM. tSHSL is (SPI_SMEM_CS_HOLD_DELAY[5:0] + 1) MSPI core clock cycles.

SPI_SMEM_CS_HOLD_R

Field SPI_SMEM_CS_HOLD reader - For SPI0 and SPI1, spi cs keep low when spi is in done phase. 1: enable 0: disable.

SPI_SMEM_CS_HOLD_TIME_R

Field SPI_SMEM_CS_HOLD_TIME reader - For SPI0 and SPI1, spi cs signal is delayed to inactive by spi clock this bits are combined with spi_mem_cs_hold bit.

SPI_SMEM_CS_SETUP_R

Field SPI_SMEM_CS_SETUP reader - For SPI0 and SPI1, spi cs is enable when spi is in prepare phase. 1: enable 0: disable.

SPI_SMEM_CS_SETUP_TIME_R

Field SPI_SMEM_CS_SETUP_TIME reader - For spi0, (cycles-1) of prepare phase by spi clock this bits are combined with spi_mem_cs_setup bit.

SPI_SMEM_ECC_16TO18_BYTE_EN_R

Field SPI_SMEM_ECC_16TO18_BYTE_EN reader - Set this bit to enable SPI0 and SPI1 ECC 16 bytes data with 2 ECC bytes mode when accesses external RAM.

SPI_SMEM_ECC_CS_HOLD_TIME_R

Field SPI_SMEM_ECC_CS_HOLD_TIME reader - SPI_SMEM_CS_HOLD_TIME + SPI_SMEM_ECC_CS_HOLD_TIME is the SPI0 and SPI1 CS hold cycles in ECC mode when accessed external RAM.

SPI_SMEM_ECC_SKIP_PAGE_CORNER_R

Field SPI_SMEM_ECC_SKIP_PAGE_CORNER reader - 1: SPI0 skips page corner when accesses external RAM. 0: Not skip page corner when accesses external RAM.

SPI_SMEM_SPLIT_TRANS_EN_R

Field SPI_SMEM_SPLIT_TRANS_EN reader - Set this bit to enable SPI0 split one AXI accesses EXT_RAM transfer into two SPI transfers when one transfer will cross flash/EXT_RAM page corner, valid no matter whether there is an ECC region or not.