Type Alias esp32c6::spi1::spi_mem_cmd::R

pub type R = R<SPI_MEM_CMD_SPEC>;

Register SPI_MEM_CMD reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn spi_mem_mst_st(&self) -> SPI_MEM_MST_ST_R

Bits 0:3 - The current status of SPI1 master FSM.

pub fn spi_mem_slv_st(&self) -> SPI_MEM_SLV_ST_R

Bits 4:7 - The current status of SPI1 slave FSM: mspi_st. 0: idle state, 1: preparation state, 2: send command state, 3: send address state, 4: wait state, 5: read data state, 6:write data state, 7: done state, 8: read data end state.

pub fn spi_mem_flash_pe(&self) -> SPI_MEM_FLASH_PE_R

Bit 17 - In user mode, it is set to indicate that program/erase operation will be triggered. The bit is combined with spi_mem_usr bit. The bit will be cleared once the operation done.1: enable 0: disable.

pub fn spi_mem_usr(&self) -> SPI_MEM_USR_R

Bit 18 - User define command enable. An operation will be triggered when the bit is set. The bit will be cleared once the operation done.1: enable 0: disable.

pub fn spi_mem_flash_hpm(&self) -> SPI_MEM_FLASH_HPM_R

Bit 19 - Drive Flash into high performance mode. The bit will be cleared once the operation done.1: enable 0: disable.

pub fn spi_mem_flash_res(&self) -> SPI_MEM_FLASH_RES_R

Bit 20 - This bit combined with reg_resandres bit releases Flash from the power-down state or high performance mode and obtains the devices ID. The bit will be cleared once the operation done.1: enable 0: disable.

pub fn spi_mem_flash_dp(&self) -> SPI_MEM_FLASH_DP_R

Bit 21 - Drive Flash into power down. An operation will be triggered when the bit is set. The bit will be cleared once the operation done.1: enable 0: disable.

pub fn spi_mem_flash_ce(&self) -> SPI_MEM_FLASH_CE_R

Bit 22 - Chip erase enable. Chip erase operation will be triggered when the bit is set. The bit will be cleared once the operation done.1: enable 0: disable.

pub fn spi_mem_flash_be(&self) -> SPI_MEM_FLASH_BE_R

Bit 23 - Block erase enable(32KB) . Block erase operation will be triggered when the bit is set. The bit will be cleared once the operation done.1: enable 0: disable.

pub fn spi_mem_flash_se(&self) -> SPI_MEM_FLASH_SE_R

Bit 24 - Sector erase enable(4KB). Sector erase operation will be triggered when the bit is set. The bit will be cleared once the operation done.1: enable 0: disable.

pub fn spi_mem_flash_pp(&self) -> SPI_MEM_FLASH_PP_R

Bit 25 - Page program enable(1 byte ~256 bytes data to be programmed). Page program operation will be triggered when the bit is set. The bit will be cleared once the operation done .1: enable 0: disable.

pub fn spi_mem_flash_wrsr(&self) -> SPI_MEM_FLASH_WRSR_R

Bit 26 - Write status register enable. Write status operation will be triggered when the bit is set. The bit will be cleared once the operation done.1: enable 0: disable.

pub fn spi_mem_flash_rdsr(&self) -> SPI_MEM_FLASH_RDSR_R

Bit 27 - Read status register-1. Read status operation will be triggered when the bit is set. The bit will be cleared once the operation done.1: enable 0: disable.

pub fn spi_mem_flash_rdid(&self) -> SPI_MEM_FLASH_RDID_R

Bit 28 - Read JEDEC ID . Read ID command will be sent when the bit is set. The bit will be cleared once the operation done. 1: enable 0: disable.

pub fn spi_mem_flash_wrdi(&self) -> SPI_MEM_FLASH_WRDI_R

Bit 29 - Write flash disable. Write disable command will be sent when the bit is set. The bit will be cleared once the operation done. 1: enable 0: disable.

pub fn spi_mem_flash_wren(&self) -> SPI_MEM_FLASH_WREN_R

Bit 30 - Write flash enable. Write enable command will be sent when the bit is set. The bit will be cleared once the operation done. 1: enable 0: disable.

pub fn spi_mem_flash_read(&self) -> SPI_MEM_FLASH_READ_R

Bit 31 - Read flash enable. Read flash operation will be triggered when the bit is set. The bit will be cleared once the operation done. 1: enable 0: disable.