Type Alias esp32c6::spi1::spi_mem_cmd::W

pub type W = W<SPI_MEM_CMD_SPEC>;

Register SPI_MEM_CMD writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn spi_mem_flash_pe(&mut self) -> SPI_MEM_FLASH_PE_W<'_, SPI_MEM_CMD_SPEC>

Bit 17 - In user mode, it is set to indicate that program/erase operation will be triggered. The bit is combined with spi_mem_usr bit. The bit will be cleared once the operation done.1: enable 0: disable.

pub fn spi_mem_usr(&mut self) -> SPI_MEM_USR_W<'_, SPI_MEM_CMD_SPEC>

Bit 18 - User define command enable. An operation will be triggered when the bit is set. The bit will be cleared once the operation done.1: enable 0: disable.

pub fn spi_mem_flash_hpm(&mut self) -> SPI_MEM_FLASH_HPM_W<'_, SPI_MEM_CMD_SPEC>

Bit 19 - Drive Flash into high performance mode. The bit will be cleared once the operation done.1: enable 0: disable.

pub fn spi_mem_flash_res(&mut self) -> SPI_MEM_FLASH_RES_W<'_, SPI_MEM_CMD_SPEC>

Bit 20 - This bit combined with reg_resandres bit releases Flash from the power-down state or high performance mode and obtains the devices ID. The bit will be cleared once the operation done.1: enable 0: disable.

pub fn spi_mem_flash_dp(&mut self) -> SPI_MEM_FLASH_DP_W<'_, SPI_MEM_CMD_SPEC>

Bit 21 - Drive Flash into power down. An operation will be triggered when the bit is set. The bit will be cleared once the operation done.1: enable 0: disable.

pub fn spi_mem_flash_ce(&mut self) -> SPI_MEM_FLASH_CE_W<'_, SPI_MEM_CMD_SPEC>

Bit 22 - Chip erase enable. Chip erase operation will be triggered when the bit is set. The bit will be cleared once the operation done.1: enable 0: disable.

pub fn spi_mem_flash_be(&mut self) -> SPI_MEM_FLASH_BE_W<'_, SPI_MEM_CMD_SPEC>

Bit 23 - Block erase enable(32KB) . Block erase operation will be triggered when the bit is set. The bit will be cleared once the operation done.1: enable 0: disable.

pub fn spi_mem_flash_se(&mut self) -> SPI_MEM_FLASH_SE_W<'_, SPI_MEM_CMD_SPEC>

Bit 24 - Sector erase enable(4KB). Sector erase operation will be triggered when the bit is set. The bit will be cleared once the operation done.1: enable 0: disable.

pub fn spi_mem_flash_pp(&mut self) -> SPI_MEM_FLASH_PP_W<'_, SPI_MEM_CMD_SPEC>

Bit 25 - Page program enable(1 byte ~256 bytes data to be programmed). Page program operation will be triggered when the bit is set. The bit will be cleared once the operation done .1: enable 0: disable.

pub fn spi_mem_flash_wrsr( &mut self ) -> SPI_MEM_FLASH_WRSR_W<'_, SPI_MEM_CMD_SPEC>

Bit 26 - Write status register enable. Write status operation will be triggered when the bit is set. The bit will be cleared once the operation done.1: enable 0: disable.

pub fn spi_mem_flash_rdsr( &mut self ) -> SPI_MEM_FLASH_RDSR_W<'_, SPI_MEM_CMD_SPEC>

Bit 27 - Read status register-1. Read status operation will be triggered when the bit is set. The bit will be cleared once the operation done.1: enable 0: disable.

pub fn spi_mem_flash_rdid( &mut self ) -> SPI_MEM_FLASH_RDID_W<'_, SPI_MEM_CMD_SPEC>

Bit 28 - Read JEDEC ID . Read ID command will be sent when the bit is set. The bit will be cleared once the operation done. 1: enable 0: disable.

pub fn spi_mem_flash_wrdi( &mut self ) -> SPI_MEM_FLASH_WRDI_W<'_, SPI_MEM_CMD_SPEC>

Bit 29 - Write flash disable. Write disable command will be sent when the bit is set. The bit will be cleared once the operation done. 1: enable 0: disable.

pub fn spi_mem_flash_wren( &mut self ) -> SPI_MEM_FLASH_WREN_W<'_, SPI_MEM_CMD_SPEC>

Bit 30 - Write flash enable. Write enable command will be sent when the bit is set. The bit will be cleared once the operation done. 1: enable 0: disable.

pub fn spi_mem_flash_read( &mut self ) -> SPI_MEM_FLASH_READ_W<'_, SPI_MEM_CMD_SPEC>

Bit 31 - Read flash enable. Read flash operation will be triggered when the bit is set. The bit will be cleared once the operation done. 1: enable 0: disable.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual