Type Alias esp32c6::spi1::spi_mem_ctrl::W

pub type W = W<SPI_MEM_CTRL_SPEC>;

Register SPI_MEM_CTRL writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn spi_mem_fdummy_rin( &mut self ) -> SPI_MEM_FDUMMY_RIN_W<'_, SPI_MEM_CTRL_SPEC>

Bit 2 - In the dummy phase of a MSPI read data transfer when accesses to flash, the signal level of SPI bus is output by the MSPI controller.

pub fn spi_mem_fdummy_wout( &mut self ) -> SPI_MEM_FDUMMY_WOUT_W<'_, SPI_MEM_CTRL_SPEC>

Bit 3 - In the dummy phase of a MSPI write data transfer when accesses to flash, the signal level of SPI bus is output by the MSPI controller.

pub fn spi_mem_fcmd_quad( &mut self ) -> SPI_MEM_FCMD_QUAD_W<'_, SPI_MEM_CTRL_SPEC>

Bit 8 - Apply 4 signals during command phase 1:enable 0: disable

pub fn spi_mem_fastrd_mode( &mut self ) -> SPI_MEM_FASTRD_MODE_W<'_, SPI_MEM_CTRL_SPEC>

Bit 13 - This bit enable the bits: spi_mem_fread_qio, spi_mem_fread_dio, spi_mem_fread_qout and spi_mem_fread_dout. 1: enable 0: disable.

pub fn spi_mem_fread_dual( &mut self ) -> SPI_MEM_FREAD_DUAL_W<'_, SPI_MEM_CTRL_SPEC>

Bit 14 - In the read operations, read-data phase apply 2 signals. 1: enable 0: disable.

pub fn spi_mem_resandres( &mut self ) -> SPI_MEM_RESANDRES_W<'_, SPI_MEM_CTRL_SPEC>

Bit 15 - The Device ID is read out to SPI_MEM_RD_STATUS register, this bit combine with spi_mem_flash_res bit. 1: enable 0: disable.

pub fn spi_mem_q_pol(&mut self) -> SPI_MEM_Q_POL_W<'_, SPI_MEM_CTRL_SPEC>

Bit 18 - The bit is used to set MISO line polarity, 1: high 0, low

pub fn spi_mem_d_pol(&mut self) -> SPI_MEM_D_POL_W<'_, SPI_MEM_CTRL_SPEC>

Bit 19 - The bit is used to set MOSI line polarity, 1: high 0, low

pub fn spi_mem_fread_quad( &mut self ) -> SPI_MEM_FREAD_QUAD_W<'_, SPI_MEM_CTRL_SPEC>

Bit 20 - In the read operations read-data phase apply 4 signals. 1: enable 0: disable.

pub fn spi_mem_wp(&mut self) -> SPI_MEM_WP_W<'_, SPI_MEM_CTRL_SPEC>

Bit 21 - Write protect signal output when SPI is idle. 1: output high, 0: output low.

pub fn spi_mem_wrsr_2b(&mut self) -> SPI_MEM_WRSR_2B_W<'_, SPI_MEM_CTRL_SPEC>

Bit 22 - two bytes data will be written to status register when it is set. 1: enable 0: disable.

pub fn spi_mem_fread_dio( &mut self ) -> SPI_MEM_FREAD_DIO_W<'_, SPI_MEM_CTRL_SPEC>

Bit 23 - In the read operations address phase and read-data phase apply 2 signals. 1: enable 0: disable.

pub fn spi_mem_fread_qio( &mut self ) -> SPI_MEM_FREAD_QIO_W<'_, SPI_MEM_CTRL_SPEC>

Bit 24 - In the read operations address phase and read-data phase apply 4 signals. 1: enable 0: disable.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual