Type Alias esp32c6::spi1::spi_mem_flash_sus_ctrl::R

pub type R = R<SPI_MEM_FLASH_SUS_CTRL_SPEC>;

Register SPI_MEM_FLASH_SUS_CTRL reader

Aliased Type

struct R { /* private fields */ }

Implementations

impl R

pub fn spi_mem_flash_per(&self) -> SPI_MEM_FLASH_PER_R

Bit 0 - program erase resume bit, program erase suspend operation will be triggered when the bit is set. The bit will be cleared once the operation done.1: enable 0: disable.

pub fn spi_mem_flash_pes(&self) -> SPI_MEM_FLASH_PES_R

Bit 1 - program erase suspend bit, program erase suspend operation will be triggered when the bit is set. The bit will be cleared once the operation done.1: enable 0: disable.

pub fn spi_mem_flash_per_wait_en(&self) -> SPI_MEM_FLASH_PER_WAIT_EN_R

Bit 2 - 1: SPI1 waits (SPI_MEM_CS_HOLD_DELAY_RES[9:0] * 4 or *128) SPI_CLK cycles after program erase resume command is sent. 0: SPI1 does not wait after program erase resume command is sent.

pub fn spi_mem_flash_pes_wait_en(&self) -> SPI_MEM_FLASH_PES_WAIT_EN_R

Bit 3 - 1: SPI1 waits (SPI_MEM_CS_HOLD_DELAY_RES[9:0] * 4 or *128) SPI_CLK cycles after program erase suspend command is sent. 0: SPI1 does not wait after program erase suspend command is sent.

pub fn spi_mem_pes_per_en(&self) -> SPI_MEM_PES_PER_EN_R

Bit 4 - Set this bit to enable PES end triggers PER transfer option. If this bit is 0, application should send PER after PES is done.

pub fn spi_mem_flash_pes_en(&self) -> SPI_MEM_FLASH_PES_EN_R

Bit 5 - Set this bit to enable Auto-suspending function.

pub fn spi_mem_pesr_end_msk(&self) -> SPI_MEM_PESR_END_MSK_R

Bits 6:21 - The mask value when check SUS/SUS1/SUS2 status bit. If the read status value is status_in[15:0](only status_in[7:0] is valid when only one byte of data is read out, status_in[15:0] is valid when two bytes of data are read out), SUS/SUS1/SUS2 = status_in[15:0]^ SPI_MEM_PESR_END_MSK[15:0].

pub fn spi_fmem_rd_sus_2b(&self) -> SPI_FMEM_RD_SUS_2B_R

Bit 22 - 1: Read two bytes when check flash SUS/SUS1/SUS2 status bit. 0: Read one byte when check flash SUS/SUS1/SUS2 status bit

pub fn spi_mem_per_end_en(&self) -> SPI_MEM_PER_END_EN_R

Bit 23 - 1: Both WIP and SUS/SUS1/SUS2 bits should be checked to insure the resume status of flash. 0: Only need to check WIP is 0.

pub fn spi_mem_pes_end_en(&self) -> SPI_MEM_PES_END_EN_R

Bit 24 - 1: Both WIP and SUS/SUS1/SUS2 bits should be checked to insure the suspend status of flash. 0: Only need to check WIP is 0.

pub fn spi_mem_sus_timeout_cnt(&self) -> SPI_MEM_SUS_TIMEOUT_CNT_R

Bits 25:31 - When SPI1 checks SUS/SUS1/SUS2 bits fail for SPI_MEM_SUS_TIMEOUT_CNT[6:0] times, it will be treated as check pass.