Module esp32c6::trace::trigger

trigger register

Structs

TRIGGER_SPEC

trigger register

Type Aliases

MEM_LOOP_R

Field MEM_LOOP reader - if this reg is 1, trace will loop wrtie trace_mem. If is 0, when mem_current_addr at mem_end_addr, it will stop at the mem_end_addr

MEM_LOOP_W

Field MEM_LOOP writer - if this reg is 1, trace will loop wrtie trace_mem. If is 0, when mem_current_addr at mem_end_addr, it will stop at the mem_end_addr

OFF_W

Field OFF writer - set 1 stop trace.

ON_W

Field ON writer - 0] set 1 start trace.

R

Register TRIGGER reader

RESTART_ENA_R

Field RESTART_ENA reader - enable encoder auto-restart, when lost package, the encoder will end, if enable auto-restart, when fifo empty, encoder will restart and send a sync package.

RESTART_ENA_W

Field RESTART_ENA writer - enable encoder auto-restart, when lost package, the encoder will end, if enable auto-restart, when fifo empty, encoder will restart and send a sync package.

W

Register TRIGGER writer