Type Alias esp32c6::uart0::conf1::W

pub type W = W<CONF1_SPEC>;

Register CONF1 writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn rxfifo_full_thrhd(&mut self) -> RXFIFO_FULL_THRHD_W<'_, CONF1_SPEC>

Bits 0:7 - It will produce rxfifo_full_int interrupt when receiver receives more data than this register value.

pub fn txfifo_empty_thrhd(&mut self) -> TXFIFO_EMPTY_THRHD_W<'_, CONF1_SPEC>

Bits 8:15 - It will produce txfifo_empty_int interrupt when the data amount in Tx-FIFO is less than this register value.

pub fn cts_inv(&mut self) -> CTS_INV_W<'_, CONF1_SPEC>

Bit 16 - Set this bit to inverse the level value of uart cts signal.

pub fn dsr_inv(&mut self) -> DSR_INV_W<'_, CONF1_SPEC>

Bit 17 - Set this bit to inverse the level value of uart dsr signal.

pub fn rts_inv(&mut self) -> RTS_INV_W<'_, CONF1_SPEC>

Bit 18 - Set this bit to inverse the level value of uart rts signal.

pub fn dtr_inv(&mut self) -> DTR_INV_W<'_, CONF1_SPEC>

Bit 19 - Set this bit to inverse the level value of uart dtr signal.

pub fn sw_dtr(&mut self) -> SW_DTR_W<'_, CONF1_SPEC>

Bit 20 - This register is used to configure the software dtr signal which is used in software flow control.

pub fn clk_en(&mut self) -> CLK_EN_W<'_, CONF1_SPEC>

Bit 21 - 1’h1: Force clock on for register. 1’h0: Support clock only when application writes registers.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual