Type Alias esp32c6::uhci0::conf1::W

pub type W = W<CONF1_SPEC>;

Register CONF1 writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn check_sum_en(&mut self) -> CHECK_SUM_EN_W<'_, CONF1_SPEC>

Bit 0 - Set this bit to enable head checksum check when receiving.

pub fn check_seq_en(&mut self) -> CHECK_SEQ_EN_W<'_, CONF1_SPEC>

Bit 1 - Set this bit to enable sequence number check when receiving.

pub fn crc_disable(&mut self) -> CRC_DISABLE_W<'_, CONF1_SPEC>

Bit 2 - Set this bit to support CRC calculation, and data integrity check bit should 1.

pub fn save_head(&mut self) -> SAVE_HEAD_W<'_, CONF1_SPEC>

Bit 3 - Set this bit to save data packet head when UHCI receive data.

pub fn tx_check_sum_re(&mut self) -> TX_CHECK_SUM_RE_W<'_, CONF1_SPEC>

Bit 4 - Set this bit to encode data packet with checksum.

pub fn tx_ack_num_re(&mut self) -> TX_ACK_NUM_RE_W<'_, CONF1_SPEC>

Bit 5 - Set this bit to encode data packet with ACK when reliable data packet is ready.

pub fn wait_sw_start(&mut self) -> WAIT_SW_START_W<'_, CONF1_SPEC>

Bit 7 - Set this bit to enable UHCI encoder transfer to ST_SW_WAIT status.

pub fn sw_start(&mut self) -> SW_START_W<'_, CONF1_SPEC>

Bit 8 - Set this bit to transmit data packet if UCHI_ENCODE_STATE is ST_SW_WAIT.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual