Type Alias esp32c6::uhci0::int_clr::W

pub type W = W<INT_CLR_SPEC>;

Register INT_CLR writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn rx_start_int_clr(&mut self) -> RX_START_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 0 - Set this bit to clear the raw interrupt of UHCI_RX_START_INT.

pub fn tx_start_int_clr(&mut self) -> TX_START_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 1 - Set this bit to clear the raw interrupt of UHCI_TX_START_INT.

pub fn rx_hung_int_clr(&mut self) -> RX_HUNG_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 2 - Set this bit to clear the raw interrupt of UHCI_RX_HUNG_INT.

pub fn tx_hung_int_clr(&mut self) -> TX_HUNG_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 3 - Set this bit to clear the raw interrupt of UHCI_TX_HUNG_INT.

pub fn send_s_reg_q_int_clr( &mut self ) -> SEND_S_REG_Q_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 4 - Set this bit to clear the raw interrupt of UHCI_SEND_S_REG_Q_INT.

pub fn send_a_reg_q_int_clr( &mut self ) -> SEND_A_REG_Q_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 5 - Set this bit to clear the raw interrupt of UHCI_SEND_A_REG_Q_INT.

pub fn outlink_eof_err_int_clr( &mut self ) -> OUTLINK_EOF_ERR_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 6 - Set this bit to clear the raw interrupt of UHCI_OUT_EOF_INT.

pub fn app_ctrl0_int_clr(&mut self) -> APP_CTRL0_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 7 - Set this bit to clear the raw interrupt of UHCI_APP_CTRL0_INT.

pub fn app_ctrl1_int_clr(&mut self) -> APP_CTRL1_INT_CLR_W<'_, INT_CLR_SPEC>

Bit 8 - Set this bit to clear the raw interrupt of UHCI_APP_CTRL1_INT.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual