Module esp32c6::uhci0::int_raw

UHCI Interrupt Raw Register

Structs

INT_RAW_SPEC

UHCI Interrupt Raw Register

Type Aliases

APP_CTRL0_INT_RAW_R

Field APP_CTRL0_INT_RAW reader - Indicates the raw interrupt of UHCI_APP_CTRL0_INT. Interrupt will be triggered when UHCI_APP_CTRL0_IN_SET is set to 1.

APP_CTRL0_INT_RAW_W

Field APP_CTRL0_INT_RAW writer - Indicates the raw interrupt of UHCI_APP_CTRL0_INT. Interrupt will be triggered when UHCI_APP_CTRL0_IN_SET is set to 1.

APP_CTRL1_INT_RAW_R

Field APP_CTRL1_INT_RAW reader - Indicates the raw interrupt of UHCI_APP_CTRL1_INT. Interrupt will be triggered when UHCI_APP_CTRL1_IN_SET is set to 1.

APP_CTRL1_INT_RAW_W

Field APP_CTRL1_INT_RAW writer - Indicates the raw interrupt of UHCI_APP_CTRL1_INT. Interrupt will be triggered when UHCI_APP_CTRL1_IN_SET is set to 1.

OUT_EOF_INT_RAW_R

Field OUT_EOF_INT_RAW reader - Indicates the raw interrupt of UHCI_OUT_EOF_INT. Interrupt will be triggered when there are errors in EOF.

OUT_EOF_INT_RAW_W

Field OUT_EOF_INT_RAW writer - Indicates the raw interrupt of UHCI_OUT_EOF_INT. Interrupt will be triggered when there are errors in EOF.

R

Register INT_RAW reader

RX_HUNG_INT_RAW_R

Field RX_HUNG_INT_RAW reader - Indicates the raw interrupt of UHCI_RX_HUNG_INT. Interrupt will be triggered when the required time of DMA receiving data exceeds the configuration value.

RX_HUNG_INT_RAW_W

Field RX_HUNG_INT_RAW writer - Indicates the raw interrupt of UHCI_RX_HUNG_INT. Interrupt will be triggered when the required time of DMA receiving data exceeds the configuration value.

RX_START_INT_RAW_R

Field RX_START_INT_RAW reader - Indicates the raw interrupt of UHCI_RX_START_INT. Interrupt will be triggered when delimiter is sent successfully.

RX_START_INT_RAW_W

Field RX_START_INT_RAW writer - Indicates the raw interrupt of UHCI_RX_START_INT. Interrupt will be triggered when delimiter is sent successfully.

SEND_A_REG_Q_INT_RAW_R

Field SEND_A_REG_Q_INT_RAW reader - Indicates the raw interrupt of UHCI_SEND_A_REG_Q_INT. Interrupt will be triggered when UHCI sends short packet successfully with always_send mode.

SEND_A_REG_Q_INT_RAW_W

Field SEND_A_REG_Q_INT_RAW writer - Indicates the raw interrupt of UHCI_SEND_A_REG_Q_INT. Interrupt will be triggered when UHCI sends short packet successfully with always_send mode.

SEND_S_REG_Q_INT_RAW_R

Field SEND_S_REG_Q_INT_RAW reader - Indicates the raw interrupt of UHCI_SEND_S_REG_Q_INT. Interrupt will be triggered when UHCI sends short packet successfully with single_send mode.

SEND_S_REG_Q_INT_RAW_W

Field SEND_S_REG_Q_INT_RAW writer - Indicates the raw interrupt of UHCI_SEND_S_REG_Q_INT. Interrupt will be triggered when UHCI sends short packet successfully with single_send mode.

TX_HUNG_INT_RAW_R

Field TX_HUNG_INT_RAW reader - Indicates the raw interrupt of UHCI_TX_HUNG_INT. Interrupt will be triggered when the required time of DMA reading RAM data exceeds the configuration value.

TX_HUNG_INT_RAW_W

Field TX_HUNG_INT_RAW writer - Indicates the raw interrupt of UHCI_TX_HUNG_INT. Interrupt will be triggered when the required time of DMA reading RAM data exceeds the configuration value.

TX_START_INT_RAW_R

Field TX_START_INT_RAW reader - Indicates the raw interrupt of UHCI_TX_START_INT. Interrupt will be triggered when DMA detects delimiter.

TX_START_INT_RAW_W

Field TX_START_INT_RAW writer - Indicates the raw interrupt of UHCI_TX_START_INT. Interrupt will be triggered when DMA detects delimiter.

W

Register INT_RAW writer