Type Alias esp32c6::uhci0::int_raw::W

pub type W = W<INT_RAW_SPEC>;

Register INT_RAW writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn rx_start_int_raw(&mut self) -> RX_START_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 0 - Indicates the raw interrupt of UHCI_RX_START_INT. Interrupt will be triggered when delimiter is sent successfully.

pub fn tx_start_int_raw(&mut self) -> TX_START_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 1 - Indicates the raw interrupt of UHCI_TX_START_INT. Interrupt will be triggered when DMA detects delimiter.

pub fn rx_hung_int_raw(&mut self) -> RX_HUNG_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 2 - Indicates the raw interrupt of UHCI_RX_HUNG_INT. Interrupt will be triggered when the required time of DMA receiving data exceeds the configuration value.

pub fn tx_hung_int_raw(&mut self) -> TX_HUNG_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 3 - Indicates the raw interrupt of UHCI_TX_HUNG_INT. Interrupt will be triggered when the required time of DMA reading RAM data exceeds the configuration value.

pub fn send_s_reg_q_int_raw( &mut self ) -> SEND_S_REG_Q_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 4 - Indicates the raw interrupt of UHCI_SEND_S_REG_Q_INT. Interrupt will be triggered when UHCI sends short packet successfully with single_send mode.

pub fn send_a_reg_q_int_raw( &mut self ) -> SEND_A_REG_Q_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 5 - Indicates the raw interrupt of UHCI_SEND_A_REG_Q_INT. Interrupt will be triggered when UHCI sends short packet successfully with always_send mode.

pub fn out_eof_int_raw(&mut self) -> OUT_EOF_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 6 - Indicates the raw interrupt of UHCI_OUT_EOF_INT. Interrupt will be triggered when there are errors in EOF.

pub fn app_ctrl0_int_raw(&mut self) -> APP_CTRL0_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 7 - Indicates the raw interrupt of UHCI_APP_CTRL0_INT. Interrupt will be triggered when UHCI_APP_CTRL0_IN_SET is set to 1.

pub fn app_ctrl1_int_raw(&mut self) -> APP_CTRL1_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 8 - Indicates the raw interrupt of UHCI_APP_CTRL1_INT. Interrupt will be triggered when UHCI_APP_CTRL1_IN_SET is set to 1.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual