Module esp32c6::usb_device::int_raw

Interrupt raw status register.

Structs

INT_RAW_SPEC

Interrupt raw status register.

Type Aliases

CRC5_ERR_INT_RAW_R

Field CRC5_ERR_INT_RAW reader - The raw interrupt bit turns to high level when CRC5 error is detected.

CRC5_ERR_INT_RAW_W

Field CRC5_ERR_INT_RAW writer - The raw interrupt bit turns to high level when CRC5 error is detected.

CRC16_ERR_INT_RAW_R

Field CRC16_ERR_INT_RAW reader - The raw interrupt bit turns to high level when CRC16 error is detected.

CRC16_ERR_INT_RAW_W

Field CRC16_ERR_INT_RAW writer - The raw interrupt bit turns to high level when CRC16 error is detected.

DTR_CHG_INT_RAW_R

Field DTR_CHG_INT_RAW reader - The raw interrupt bit turns to high level when level of DTR from usb serial channel is changed.

DTR_CHG_INT_RAW_W

Field DTR_CHG_INT_RAW writer - The raw interrupt bit turns to high level when level of DTR from usb serial channel is changed.

GET_LINE_CODE_INT_RAW_R

Field GET_LINE_CODE_INT_RAW reader - The raw interrupt bit turns to high level when level of GET LINE CODING request is received.

GET_LINE_CODE_INT_RAW_W

Field GET_LINE_CODE_INT_RAW writer - The raw interrupt bit turns to high level when level of GET LINE CODING request is received.

IN_TOKEN_REC_IN_EP1_INT_RAW_R

Field IN_TOKEN_REC_IN_EP1_INT_RAW reader - The raw interrupt bit turns to high level when IN token for IN endpoint 1 is received.

IN_TOKEN_REC_IN_EP1_INT_RAW_W

Field IN_TOKEN_REC_IN_EP1_INT_RAW writer - The raw interrupt bit turns to high level when IN token for IN endpoint 1 is received.

JTAG_IN_FLUSH_INT_RAW_R

Field JTAG_IN_FLUSH_INT_RAW reader - The raw interrupt bit turns to high level when flush cmd is received for IN endpoint 2 of JTAG.

JTAG_IN_FLUSH_INT_RAW_W

Field JTAG_IN_FLUSH_INT_RAW writer - The raw interrupt bit turns to high level when flush cmd is received for IN endpoint 2 of JTAG.

OUT_EP1_ZERO_PAYLOAD_INT_RAW_R

Field OUT_EP1_ZERO_PAYLOAD_INT_RAW reader - The raw interrupt bit turns to high level when OUT endpoint 1 received packet with zero palyload.

OUT_EP1_ZERO_PAYLOAD_INT_RAW_W

Field OUT_EP1_ZERO_PAYLOAD_INT_RAW writer - The raw interrupt bit turns to high level when OUT endpoint 1 received packet with zero palyload.

OUT_EP2_ZERO_PAYLOAD_INT_RAW_R

Field OUT_EP2_ZERO_PAYLOAD_INT_RAW reader - The raw interrupt bit turns to high level when OUT endpoint 2 received packet with zero palyload.

OUT_EP2_ZERO_PAYLOAD_INT_RAW_W

Field OUT_EP2_ZERO_PAYLOAD_INT_RAW writer - The raw interrupt bit turns to high level when OUT endpoint 2 received packet with zero palyload.

PID_ERR_INT_RAW_R

Field PID_ERR_INT_RAW reader - The raw interrupt bit turns to high level when pid error is detected.

PID_ERR_INT_RAW_W

Field PID_ERR_INT_RAW writer - The raw interrupt bit turns to high level when pid error is detected.

R

Register INT_RAW reader

RTS_CHG_INT_RAW_R

Field RTS_CHG_INT_RAW reader - The raw interrupt bit turns to high level when level of RTS from usb serial channel is changed.

RTS_CHG_INT_RAW_W

Field RTS_CHG_INT_RAW writer - The raw interrupt bit turns to high level when level of RTS from usb serial channel is changed.

SERIAL_IN_EMPTY_INT_RAW_R

Field SERIAL_IN_EMPTY_INT_RAW reader - The raw interrupt bit turns to high level when Serial Port IN Endpoint is empty.

SERIAL_IN_EMPTY_INT_RAW_W

Field SERIAL_IN_EMPTY_INT_RAW writer - The raw interrupt bit turns to high level when Serial Port IN Endpoint is empty.

SERIAL_OUT_RECV_PKT_INT_RAW_R

Field SERIAL_OUT_RECV_PKT_INT_RAW reader - The raw interrupt bit turns to high level when Serial Port OUT Endpoint received one packet.

SERIAL_OUT_RECV_PKT_INT_RAW_W

Field SERIAL_OUT_RECV_PKT_INT_RAW writer - The raw interrupt bit turns to high level when Serial Port OUT Endpoint received one packet.

SET_LINE_CODE_INT_RAW_R

Field SET_LINE_CODE_INT_RAW reader - The raw interrupt bit turns to high level when level of SET LINE CODING request is received.

SET_LINE_CODE_INT_RAW_W

Field SET_LINE_CODE_INT_RAW writer - The raw interrupt bit turns to high level when level of SET LINE CODING request is received.

SOF_INT_RAW_R

Field SOF_INT_RAW reader - The raw interrupt bit turns to high level when SOF frame is received.

SOF_INT_RAW_W

Field SOF_INT_RAW writer - The raw interrupt bit turns to high level when SOF frame is received.

STUFF_ERR_INT_RAW_R

Field STUFF_ERR_INT_RAW reader - The raw interrupt bit turns to high level when stuff error is detected.

STUFF_ERR_INT_RAW_W

Field STUFF_ERR_INT_RAW writer - The raw interrupt bit turns to high level when stuff error is detected.

USB_BUS_RESET_INT_RAW_R

Field USB_BUS_RESET_INT_RAW reader - The raw interrupt bit turns to high level when usb bus reset is detected.

USB_BUS_RESET_INT_RAW_W

Field USB_BUS_RESET_INT_RAW writer - The raw interrupt bit turns to high level when usb bus reset is detected.

W

Register INT_RAW writer