Type Alias esp32c6::usb_device::int_raw::W

pub type W = W<INT_RAW_SPEC>;

Register INT_RAW writer

Aliased Type

struct W { /* private fields */ }

Implementations

impl W

pub fn jtag_in_flush_int_raw( &mut self ) -> JTAG_IN_FLUSH_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 0 - The raw interrupt bit turns to high level when flush cmd is received for IN endpoint 2 of JTAG.

pub fn sof_int_raw(&mut self) -> SOF_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 1 - The raw interrupt bit turns to high level when SOF frame is received.

pub fn serial_out_recv_pkt_int_raw( &mut self ) -> SERIAL_OUT_RECV_PKT_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 2 - The raw interrupt bit turns to high level when Serial Port OUT Endpoint received one packet.

pub fn serial_in_empty_int_raw( &mut self ) -> SERIAL_IN_EMPTY_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 3 - The raw interrupt bit turns to high level when Serial Port IN Endpoint is empty.

pub fn pid_err_int_raw(&mut self) -> PID_ERR_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 4 - The raw interrupt bit turns to high level when pid error is detected.

pub fn crc5_err_int_raw(&mut self) -> CRC5_ERR_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 5 - The raw interrupt bit turns to high level when CRC5 error is detected.

pub fn crc16_err_int_raw(&mut self) -> CRC16_ERR_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 6 - The raw interrupt bit turns to high level when CRC16 error is detected.

pub fn stuff_err_int_raw(&mut self) -> STUFF_ERR_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 7 - The raw interrupt bit turns to high level when stuff error is detected.

pub fn in_token_rec_in_ep1_int_raw( &mut self ) -> IN_TOKEN_REC_IN_EP1_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 8 - The raw interrupt bit turns to high level when IN token for IN endpoint 1 is received.

pub fn usb_bus_reset_int_raw( &mut self ) -> USB_BUS_RESET_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 9 - The raw interrupt bit turns to high level when usb bus reset is detected.

pub fn out_ep1_zero_payload_int_raw( &mut self ) -> OUT_EP1_ZERO_PAYLOAD_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 10 - The raw interrupt bit turns to high level when OUT endpoint 1 received packet with zero palyload.

pub fn out_ep2_zero_payload_int_raw( &mut self ) -> OUT_EP2_ZERO_PAYLOAD_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 11 - The raw interrupt bit turns to high level when OUT endpoint 2 received packet with zero palyload.

pub fn rts_chg_int_raw(&mut self) -> RTS_CHG_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 12 - The raw interrupt bit turns to high level when level of RTS from usb serial channel is changed.

pub fn dtr_chg_int_raw(&mut self) -> DTR_CHG_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 13 - The raw interrupt bit turns to high level when level of DTR from usb serial channel is changed.

pub fn get_line_code_int_raw( &mut self ) -> GET_LINE_CODE_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 14 - The raw interrupt bit turns to high level when level of GET LINE CODING request is received.

pub fn set_line_code_int_raw( &mut self ) -> SET_LINE_CODE_INT_RAW_W<'_, INT_RAW_SPEC>

Bit 15 - The raw interrupt bit turns to high level when level of SET LINE CODING request is received.

pub unsafe fn bits(&mut self, bits: u32) -> &mut Self

Writes raw bits to the register.

Safety

Passing incorrect value can cause undefined behaviour. See reference manual